+++ bin/botanix-up/Cargo.toml
@@ -14,6 +14,7 @@ path = "src/main.rs"
 
 [dependencies]
 anyhow = "1.0.95"
+botanix-cli-args = { workspace = true }
 clap = { version = "4.5", features = ["derive"] }
 hex = "0.4.3"
 rand = { workspace = true }
