+++ bin/test-suite/Cargo.toml
@@ -28,10 +28,11 @@ botanix-authority-peg = { workspace = true, features = ["secp256k1"] }
 
 # Botanix
 botanix-btc-wallet = { path = "../../crates/botanix-btc-wallet" }
+botanix-cli-args = { workspace = true }
 botanix-comet-bft-rpc = { workspace = true }
 botanix-data-parser = { workspace = true }
-btc-server = { path = "../btc-server" }
 botanix-storage = { workspace = true }
+btc-server = { path = "../btc-server" }
 
 # other dependencies
 bytes = "1.4.0"
