+++ bin/test-suite/src/suite/consensus/common/poa_node.rs
@@ -13,6 +13,7 @@ use bitcoin::hashes::Hash;
 use botanix_authority_edh::extra_data_header::{
     ExtraDataHeader, CHAIN_VERSION, EXTRA_HEADER_VERSION,
 };
+use botanix_cli_args::federation_args::{FedMemberPubKey, FederationTomlConfig};
 use botanix_storage::BotanixProviderFactory;
 use btcserverlib::extended_client::{BtcServerExtendedApi, BtcServerExtendedClient};
 use client::{Empty, GetSessionIdsRequest, GetSigningStatusRequest, SigningStatus};
@@ -20,7 +21,6 @@ use ethers::{
     providers::{Middleware, PeerInfo, StreamExt},
     types::{BlockId, BlockNumber, H256},
 };
-use reth::args::{FedMemberPubKey, FederationTomlConfig};
 use reth_chainspec::BOTANIX_TESTNET;
 use reth_db::{
     mdbx::{DatabaseArguments, MaxReadTransactionDuration},
