+++ crates/botanix-utils/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "botanix-utils"
+description = "Utils for Botanix"
+version = { workspace = true }
+edition = { workspace = true }
+rust-version = { workspace = true }
+license = { workspace = true }
+homepage = { workspace = true }
+repository = { workspace = true }
+
+[dependencies]
+tracing = { workspace = true }
+
+[features]
+default = []
