+++ crates/botanix-utils/src/panic_hook.rs
@@ -1,0 +1,23 @@
+use tracing::error;
+
+/// Adds a panic hook to log the panic information
+pub fn set_panic_hook() {
+    std::panic::set_hook(Box::new(|panic_info| {
+        let payload = panic_info.payload();
+
+        #[allow(clippy::manual_map)]
+        let payload = if let Some(s) = payload.downcast_ref::<&str>() {
+            Some(&**s)
+        } else if let Some(s) = payload.downcast_ref::<String>() {
+            Some(s.as_str())
+        } else {
+            None
+        };
+
+        let location = panic_info.location().map(|l| l.to_string());
+
+        error!(panic.payload = payload, panic.location = location, "Uncaught panic");
+
+        std::process::exit(1);
+    }));
+}
