+++ crates/consensus/authority/Cargo.toml
@@ -12,7 +12,6 @@ description = "A PoA consensus engine"
 [dependencies]
 alloy-rpc-types-engine = { version = "0.2.1", default-features = false }
 async-trait = { workspace = true }
-botanix-storage = { workspace = true }
 
 ## Bitcoin
 bitcoin = { workspace = true }
@@ -24,8 +23,10 @@ botanix-authority-peg = { workspace = true, features = ["secp256k1"] }
 botanix-authority-rsp = { workspace = true }
 botanix-bitcoin-checkpoint = { workspace = true }
 botanix-btc-wallet = { workspace = true }
+botanix-cli-args = { workspace = true }
 botanix-comet-bft-rpc = { workspace = true }
 botanix-data-parser = { workspace = true }
+botanix-storage = { workspace = true }
 btc-server = { path = "../../../bin/btc-server" }
 
 # others
@@ -103,4 +104,3 @@ rand = "0.8.5"
 reth-cli-runner = { workspace = true }
 reth-db-common = { workspace = true }
 reth-evm = { workspace = true, features = ["test-utils"] }
-tempfile = { workspace = true }
