+++ crates/node/core/Cargo.toml
@@ -21,7 +21,7 @@ askama = "0.12.1"
 # bitcoin
 bitcoin = { workspace = true }
 botanix-authority-edh = { workspace = true }
-botanix-btc-wallet = { workspace = true }
+botanix-cli-args = { workspace = true }
 
 # Botanix
 btc-server = { path = "../../../bin/btc-server" }
@@ -31,9 +31,6 @@ derive_more = { workspace = true }
 
 # io
 dirs-next = "2.0.0"
-
-# displaydoc
-displaydoc = "0.2.4"
 
 # misc
 eyre = { workspace = true }
@@ -80,7 +77,6 @@ secp256k1 = { workspace = true, features = ["global-context", "rand-std", "recov
 serde = { workspace = true }
 serde_json = { workspace = true }
 shellexpand = { workspace = true }
-thiserror = { workspace = true }
 toml = { workspace = true }
 
 # tracing
