+++ crates/node/core/src/args/mod.rs
@@ -4,10 +4,6 @@
 mod network;
 pub use network::{DiscoveryArgs, NetworkArgs};
 
-/// Configuration for the genesis block (toml)
-mod federation_args;
-pub use federation_args::{FedMemberPubKey, FederationTomlConfig};
-
 /// RpcServerArg struct for configuring the RPC
 mod rpc_server;
 pub use rpc_server::RpcServerArgs;
@@ -15,10 +11,6 @@ pub use rpc_server::RpcServerArgs;
 /// `RpcStateCacheArgs` struct for configuring RPC state cache
 mod rpc_state_cache;
 pub use rpc_state_cache::RpcStateCacheArgs;
-
-/// `StateSyncArgs` struct for configuring state sync
-mod state_sync;
-pub use state_sync::StateSyncArgs;
 
 /// DebugArgs struct for debugging purposes
 mod debug;
@@ -55,14 +47,6 @@ pub use dev::DevArgs;
 /// PruneArgs for configuring the pruning and full node
 mod pruning;
 pub use pruning::PruningArgs;
-
-/// `BitcoindArgs` for configuration settings of the bitcoind instance
-mod bitcoind_args;
-pub use bitcoind_args::BitcoindArgs;
-
-/// `FrostArgs` for configuration settings of the frost protocol
-mod frost_args;
-pub use frost_args::FrostArgs;
 
 /// DatadirArgs for configuring data storage paths
 mod datadir_args;
