+++ crates/node/core/src/args/utils.rs
@@ -6,19 +6,16 @@ use bitcoin::hashes::Hash;
 use botanix_authority_edh::extra_data_header::{
     ExtraDataHeader, CHAIN_VERSION, EXTRA_HEADER_VERSION,
 };
+use botanix_cli_args::federation_args::FederationTomlConfig;
 use reth_chainspec::{
     create_botanix_config_with_genesis, BotanixMainnetGenesisConfig, BotanixTestnetGenesisConfig,
     ChainSpec, BOTANIX_MAINNET, BOTANIX_MAINNET_CHAIN_ID, BOTANIX_TESTNET,
-    BOTANIX_TESTNET_CHAIN_ID, DEV,
+    BOTANIX_TESTNET_CHAIN_ID, DEV, HOLESKY, MAINNET, SEPOLIA,
 };
 use reth_fs_util as fs;
 use reth_primitives::{constants::nums_secp256k1_pk, Address};
 use std::{path::PathBuf, str::FromStr, sync::Arc};
 use tracing::info;
-
-use reth_chainspec::{HOLESKY, MAINNET, SEPOLIA};
-
-use super::FederationTomlConfig;
 
 /// Chains supported by reth. First value should be used as the default.
 pub const SUPPORTED_CHAINS: &[&str] =
